6420 - Move wagtail hooks into wagtail_hooks.py file (#7054)
* 6420 - Move wagtail hooks into wagtail_hooks.py file

* Keep Customization section seperate

edits

Linter corrections
